fix: move select to common
